Added create expense participant list.

+++ app/src/main/java/com/lukasbrand/sharedwallet/data/ExpenseParticipant.kt
@@ -2,6 +2,6 @@ package com.lukasbrand.sharedwallet.data
 
 data class ExpenseParticipant(
     val user: User,
-    val expensePercentage: Double,
+    val expensePercentage: Int,
     val hasPaid: Boolean
 )
